Successfully integrated Regret Synthesis code with my Simulator. Added more functionlaity in the main module and the World env to help in automating simulation and human intervention.
main.py: Added function to automate the process extracting the concerned sequence of actions(controls), and executing in the file.

src/graph_construction/two_player_game.py: Fixed a minor bug where we had all the copies of the initial state of the transition system in the two player game as init state as well. AN abstraction should always containly only 1 init state.