Working illustration of modified form of diagnoal placement with 3 objects and 6 locations
Please check the corresponding pddl file and the main/py file code to reproduce the modified diagonal placmeent sim. NOTE: As long as the robot is operating in its winning region, the Sys player will explore cooperation. ANother bootleneck we are running into is scalability beyond 6 locs and 4 objs. Anymore locations or objects, we are unable to construct the transistion system.